Впринципе все изменил по правкам, кроме season_image_url и title_image_url. Проверьте Comment и Rating, и связи. Вроде все работает

+++ src/main/java/ru/youmiteru/backend/domain/Comment.java
@@ -26,5 +26,5 @@ public class Comment {
     private Double ratingValue;
 
     @Column(name = "reply_to")
-    private Double replyTo;
+    private Long replyTo;
 }
